--- packages/shared/clickhouse/seed-constants.ts
@@ -1,332 +1,0 @@
-export const LANGFUSE_SEED_TRACE_DATA = {
-  id: "d1095d74-a996-480b-9eaf-143aaf9f2609",
-  name: "LangGraph",
-  input: JSON.stringify({
-    messages: [
-      {
-        content: "what is the weather in sf",
-        additional_kwargs: {},
-        response_metadata: {},
-        type: "human",
-        name: null,
-        id: "2166b887-b9fb-4282-a9f7-828d4a6cca54",
-        example: false,
-      },
-    ],
-  }),
-  output: JSON.stringify({
-    messages: [
-      {
-        content: "what is the weather in sf",
-        additional_kwargs: {},
-        response_metadata: {},
-        type: "human",
-        name: null,
-        id: "2166b887-b9fb-4282-a9f7-828d4a6cca54",
-        example: false,
-      },
-      {
-        content: "",
-        additional_kwargs: {
-          tool_calls: [
-            {
-              id: "call_pNLR4DoZiptb5xGlb299QsoN",
-              function: {
-                arguments: JSON.stringify({
-                  query: "current weather in San Francisco",
-                }),
-                name: "search",
-              },
-              type: "function",
-            },
-          ],
-          refusal: null,
-        },
-        response_metadata: {
-          token_usage: {
-            completion_tokens: 17,
-            prompt_tokens: 48,
-            total_tokens: 65,
-          },
-          model_name: "gpt-4o-mini-2024-07-18",
-          system_fingerprint: "fp_34a54ae93c",
-          finish_reason: "tool_calls",
-        },
-        type: "ai",
-        name: null,
-        id: "run-14c4801c-ae47-4c4a-8d49-29ba08c5679d-0",
-        example: false,
-      },
-      {
-        content: "It's 60 degrees and foggy.",
-        additional_kwargs: {},
-        response_metadata: {},
-        type: "tool",
-        name: "search",
-        id: "220ba511-6816-4d1a-8acf-8e5791c2d88e",
-        tool_call_id: "call_pNLR4DoZiptb5xGlb299QsoN",
-      },
-      {
-        content:
-          "The current weather in San Francisco is 60 degrees and foggy.",
-        additional_kwargs: {
-          refusal: null,
-        },
-        response_metadata: {
-          token_usage: {
-            completion_tokens: 15,
-            prompt_tokens: 80,
-            total_tokens: 95,
-          },
-          model_name: "gpt-4o-mini-2024-07-18",
-        },
-        type: "ai",
-        name: null,
-        id: "run-f051be3b-2656-4b19-8704-a2e492306131-0",
-        example: false,
-      },
-    ],
-  }),
-  metadata: JSON.stringify({ thread_id: 42 }),
-  environment: "default",
-  tags: [] as string[],
-  bookmarked: false,
-  public: true,
-  release: null,
-  version: null,
-  userId: null,
-  sessionId: null,
-  latency: 2.756,
-};
-
-export const LANGFUSE_SEED_OBSERVATIONS = [
-  {
-    id: "901cb76a-af59-4982-b468-6e4323eb00d3",
-    type: "SPAN",
-    name: "LangGraph",
-    parentObservationId: null,
-    startTimeOffset: 0,
-    endTimeOffset: 2756,
-    metadata: JSON.stringify({}),
-    level: "DEFAULT",
-    model: null,
-    modelParameters: null,
-    usageDetails: {},
-    costDetails: {},
-    latency: 2756,
-  },
-  {
-    id: "0cdeede0-7a8d-4cbb-a4c6-7b10023fed60",
-    type: "SPAN",
-    name: "agent",
-    parentObservationId: "901cb76a-af59-4982-b468-6e4323eb00d3",
-    startTimeOffset: 4,
-    endTimeOffset: 1865,
-    metadata: JSON.stringify({}),
-    level: "DEFAULT",
-    model: null,
-    modelParameters: null,
-    usageDetails: {},
-    costDetails: {},
-    latency: 1861,
-  },
-  {
-    id: "b3662aaf-6a6c-4170-943a-6019c6528526",
-    type: "GENERATION",
-    name: "ChatOpenAI",
-    parentObservationId: "0cdeede0-7a8d-4cbb-a4c6-7b10023fed60",
-    startTimeOffset: 4,
-    endTimeOffset: 1862,
-    metadata: JSON.stringify({}),
-    level: "DEFAULT",
-    model: "gpt-4o-mini-2024-07-18",
-    modelParameters: JSON.stringify({ temperature: 0.7 }),
-    usageDetails: {
-      input: 48,
-      output: 17,
-      total: 65,
-      input_audio: 0,
-      input_cache_read: 0,
-      output_audio: 0,
-      output_reasoning: 0,
-    },
-    costDetails: {
-      input: 0.0000072,
-      output: 0.0000102,
-      input_cache_read: 0,
-      total: 0.0000174,
-    },
-    latency: 1858,
-  },
-  {
-    id: "d4932b0a-5792-461d-8dfa-7c4847fb8cb5",
-    type: "SPAN",
-    name: "tools",
-    parentObservationId: "901cb76a-af59-4982-b468-6e4323eb00d3",
-    startTimeOffset: 1866,
-    endTimeOffset: 1874,
-    metadata: JSON.stringify({}),
-    level: "DEFAULT",
-    model: null,
-    modelParameters: null,
-    usageDetails: {},
-    costDetails: {},
-    latency: 8,
-  },
-  {
-    id: "c50bf74b-d915-4a80-8039-aedd4de816e5",
-    type: "SPAN",
-    name: "search",
-    parentObservationId: "d4932b0a-5792-461d-8dfa-7c4847fb8cb5",
-    startTimeOffset: 1872,
-    endTimeOffset: 1872,
-    metadata: JSON.stringify({}),
-    level: "DEFAULT",
-    model: null,
-    modelParameters: null,
-    usageDetails: {},
-    costDetails: {},
-    latency: 0,
-  },
-  {
-    id: "486b8b8a-26e8-4303-85ae-7ec262f0f893",
-    type: "SPAN",
-    name: "agent",
-    parentObservationId: "901cb76a-af59-4982-b468-6e4323eb00d3",
-    startTimeOffset: 1874,
-    endTimeOffset: 2750,
-    metadata: JSON.stringify({}),
-    level: "DEFAULT",
-    model: null,
-    modelParameters: null,
-    usageDetails: {},
-    costDetails: {},
-    latency: 876,
-  },
-  {
-    id: "23b899dc-d5a2-41ec-ba62-ef0f991f417f",
-    type: "GENERATION",
-    name: "ChatOpenAI",
-    parentObservationId: "486b8b8a-26e8-4303-85ae-7ec262f0f893",
-    startTimeOffset: 1875,
-    endTimeOffset: 2747,
-    metadata: JSON.stringify({}),
-    level: "DEFAULT",
-    model: "gpt-4o-mini-2024-07-18",
-    modelParameters: JSON.stringify({ temperature: 0.7 }),
-    usageDetails: {
-      input: 80,
-      output: 15,
-      total: 95,
-      input_audio: 0,
-      input_cache_read: 0,
-      output_audio: 0,
-      output_reasoning: 0,
-    },
-    costDetails: {
-      input: 0.000012,
-      output: 0.000009,
-      input_cache_read: 0,
-      total: 0.000021,
-    },
-    latency: 872,
-  },
-];
-
-export const REALISTIC_TRACE_NAMES = [
-  "LangGraph",
-  "ChatCompletion",
-  "DocumentAnalysis",
-  "CodeGeneration",
-  "DataProcessing",
-  "QueryExecution",
-  "ModelInference",
-  "WebScraping",
-  "TextSummarization",
-  "ImageClassification",
-];
-
-export const REALISTIC_SPAN_NAMES = [
-  "agent",
-  "tools",
-  "search",
-  "retrieval",
-  "preprocessing",
-  "validation",
-  "transformation",
-  "classification",
-  "extraction",
-  "postprocessing",
-];
-
-export const REALISTIC_GENERATION_NAMES = [
-  "ChatOpenAI",
-  "GPT-4",
-  "Claude-3",
-  "Gemini",
-  "Llama-2",
-  "PaLM",
-  "CodeLlama",
-  "Mistral",
-  "Falcon",
-  "Vicuna",
-];
-
-export const REALISTIC_MODELS = [
-  "gpt-4o-mini-2024-07-18",
-  "gpt-4-turbo-2024-04-09",
-  "claude-3-haiku-20240307",
-  "claude-3-sonnet-20240229",
-  "claude-3-opus-20240229",
-  "gemini-pro",
-  "llama-2-70b-chat",
-  "mistral-7b-instruct",
-  "codellama-34b-instruct",
-];
-
-export const REALISTIC_USER_INPUTS = [
-  "What is the weather in San Francisco?",
-  "Summarize this document for me",
-  "Generate a Python function to sort a list",
-  "Explain quantum computing in simple terms",
-  "Analyze this data and provide insights",
-  "Translate this text to Spanish",
-  "Create a marketing email for our product",
-  "Debug this code and fix the errors",
-  "What are the latest trends in AI?",
-  "Help me plan a trip to Europe",
-];
-
-export const REALISTIC_AI_RESPONSES = [
-  "The current weather in San Francisco is 60 degrees and foggy.",
-  "Here's a summary of the key points from the document...",
-  "Here's a Python function that sorts a list efficiently...",
-  "Quantum computing uses quantum mechanics principles...",
-  "Based on the data analysis, I found the following insights...",
-  "Here's the Spanish translation of your text...",
-  "I've created a compelling marketing email for your product...",
-  "I found several issues in your code and here are the fixes...",
-  "The latest AI trends include large language models...",
-  "Here's a detailed 10-day European itinerary for you...",
-];
-
-export const REALISTIC_METADATA_EXAMPLES = [
-  { thread_id: 42, session_type: "interactive" },
-  { user_id: "user_123", conversation_id: "conv_456" },
-  { model_version: "v2.1", temperature: 0.7 },
-  { request_id: "req_789", timestamp: "2024-05-23T15:42:11.996Z" },
-  { environment: "production", region: "us-west-2" },
-  {
-    document_type: "state_of_union",
-    file_size: "142KB",
-    processing_time: "2.3s",
-  },
-  { data_source: "product_catalog", record_count: 30, format: "nested_json" },
-  { analysis_type: "sentiment", language: "en", confidence: 0.92 },
-  {
-    extraction_task: "policy_analysis",
-    domain: "politics",
-    entities_found: 15,
-  },
-  { file_type: "JSON", validation: "passed", schema_version: "v1.2" },
-];
